OCPBUGS-3761: migrate Events integration test to Cypress to mitigate flakes
